fixed warnings

+++ src/megaapi_impl.cpp
@@ -13169,7 +13169,7 @@ void MegaApiImpl::sendPendingRequests()
                 break;
             }
 
-            if (connections > MegaClient::MAX_NUM_CONNECTIONS)
+            if ((unsigned int) connections > MegaClient::MAX_NUM_CONNECTIONS)
             {
                 e = API_ETOOMANY;
                 break;
